Reload world on file change (leaky)
Right now we never free memory used by assets, entities and world data.
This is fine for now since we don't really have the issue in release builds but
at some point we are going to have to handle this correctly.
Also, assets are in the platform memory arena right now.